Compile with 16.04, link against external g2o

+++ ssa/apps/ssa_viewer_3d/standardCamera.h
@@ -8,14 +8,14 @@ class StandardCamera : public qglviewer::Camera
   public:
     StandardCamera() : qglviewer::Camera(), _standard(true), _closeRange(0.001), _farRange(1000.) {};
 
-    float zNear() const {
+    qreal zNear() const {
       if (_standard) 
         return _closeRange; 
       else 
         return Camera::zNear(); 
     }
 
-    float zFar() const
+    qreal zFar() const
     {  
       if (_standard) 
         return _farRange;
